Add ServiceRecord hook again, refactor ServiceUtils
