+++ zygote/src/main/java/org/frknkrc44/hma_oss/zygote/util/ContextUtils.kt
@@ -1,0 +1,11 @@
+package org.frknkrc44.hma_oss.zygote.util
+
+import android.app.ActivityThread
+
+object ContextUtils {
+    private val application get () = ActivityThread.currentActivityThread().application!!
+
+    val packageManager get() = application.packageManager!!
+
+    val contentResolver get() = application.contentResolver!!
+}
